Dodanie Walidatorów

+++ managing-humanitarian-collections-api/Entities/Profile.cs
@@ -10,7 +10,7 @@
         #endregion
 
         #region Organizator
-        public int? Nip { get; set; }
+        public long? Nip { get; set; }
         public int? Regon { get; set; }
         public string? Name { get; set; }
         public int UserId { get; set; }
